When passed an array of values in a hash, add each element
When receiving an array of values directly, a condition node adds each
element of the array individually to the condition's underlying values.

However, when the array of values received is nested within a hash, the
array is added all at once, as one value to the condition's underlying
values.

This changes the array-within-hash behavior to match that of the
direct-array behavior, adding each element of the array individually to
the condition's underlying values.