Fix specs broken by Rails 3 on Ruby 2.3.0.
With the new Ruby 2.3.0, Active Record 3 association conditions are not
being recognized when using hash syntax instead of SQL string literals.

Using string literals makes the spec work again, but it may be worth
checking whether it is Ransack or Active Record that is breaking with
Ruby 2.3.0 on this issue.