Silence dynamic route deprecation messages in Ransack specs
Using dynamic :controller or :action segments in routes has been a
source of a number of security issues in production and is deprecated
in Rails 5.0 since this [pull request] and is planned to be removed in
Rails 5.1.

Ransack is still maintaining compatibility with legacy Rails 3 and 4
versions, so this commit silences the deprecation messages for now
while running the test suite.

[pull request]: https://github.com/rails/rails/pull/23980