Add support for default predicates
Currently predicates are required, otherwise the conditions are ignored
or they will raise an error.
When using Ransack for filtering in json requests, it would be nice to
allow an empty predicate to equal to true.

So instead of having to explicitly pass '*_eq':

    {
      product_id_eq: 'product-1',
      type_eq:       'admin
    }

we can do the following:

    {
      product_id: 'product-1',
      type:       'admin
    }

This adds a `:default_predicate` configuration option that will be used
in case of a missing predicate.