Fix enum support (#1559)
The casting of condition values is incorrect for enums. The value is a
number but enum type under the hood (DB) is an integer, so the string
values get casted to zero.

Apparently this casting is not actually needed (except for converting
dates) because ActiveRecord handles this just fine internally. So I
removed it and it seems to fix enums and does not make any specs fail.

Co-authored-by: Peko <91131029+pekopekopekopayo@users.noreply.github.com>