Check type before sending `#value` message to predicate (#1468)
The previous implementation was giving for granted that every predicate
respond  to `#value`, but that doesn't seem to be the case at least when
having a `Arel::SelectManager`.

by simply inverting the terms of the existing AND check we can fix the issue
without introducing unknown side effects.

In order to test the change, a new ransacker has been added to the Person model.

Co-authored-by: Sean <sean@immersive-app.com>